Update Zed.py
Rewrote the ENTIRE bot, refactoring it to better code standards. Added autocomplete functionality, more character traits in the database, and added the MinacX condition support. This is likely only a handful of features implemented in this commit; individual commits were given in the 'refactoring' branch, which will now be removed.