Update Zed.py
Deprecated the CSV database system and converted to SQL.
CSVs are now only used to set up the SQLite database when the bot launches, this allows me to still develop with the speed and ease of a CSV-based database but also the computational speed of an SQL database. Once development has gotten closer to the final product (with more bugs found and squashed) the CSVs can be removed entirely after one final conversion.