Update Zed.py
Updated the Database creation, adding userIDs that are tied to a user's character. Also adjusted the autocomplete function for characters to only show characters for users in the guild that the interaction is being used in... that makes sense, right? Right.
Also added many ideas to build upon this, which will be next.
Also, I fixed a bug in /roll_ability :)